Improve reporting.
Having compared the performance with the pytorch code this is based on,
I've concluded that the using SGD instead of Adam is most of the reason
why the current code doesn't work. However, it's also very possible that
the gradients are being calculated incorrectly.

Up next, I'm going to implement Adam, and then we'll see where that
takes us.